Change iteritems() to items() in volumes [BK-1292]
So I added a bunch of logs and figured out the cause - iteritems for MultiDict is broken.
search_opts is a MultiDict. When it has one option in it, things work fine, and it runs literal_eval on the single value.
When there are two, then iteritems() returns the first item twice, and then literal_eval isn't run on the second one.
In our case we have 2 keys/values: {"all_tenants":"1" and "id":"...".
It will run literal_eval on "1" twice, and not run it the second, so we end up with a string instead of a list.
The easy fix is to replace iteritems() with items(), and that works.